+++ socket_recvfrom.c
@@ -47,8 +47,8 @@
 
 /****************************************************************************/
 
-int
-recvfrom(int sockfd,void *buff,int len,int flags,struct sockaddr *from,socklen_t *fromlen)
+ssize_t
+recvfrom(int sockfd,void *buff,size_t len,int flags,struct sockaddr *from,socklen_t *fromlen)
 {
 	struct fd * fd;
 	int result = ERROR;
